Day 8/requirements.txt
langchain
langchain-google-genai
streamlit
python-dotenv
pypdf
sentence-transformers
faiss-cpu
sentence-transformers
